Now there won't be any redundant vertices
How I did it:

The indices array does not need to resize, it just needs to change the
indices of the redundant vertices in to the index of the first original
vertex.

The position, texcoord and normal array do need to resize. So, before
adding new vertex in the 3 arrays, I check if the same vertex already
exist. If so, do not add it., and use the index of the first original
vertex to represent the current vertex.  If not so, add it, and use a
new index to represent the current vertex.